requirements.txt
asgiref==3.8.1
better-exceptions==0.3.3
colorama==0.4.6
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
packaging==24.1
Pygments==2.18.0
PyJWT==2.8.0
rich==13.7.1
sqlparse==0.5.1
tzdata==2024.1
